python_scripts/requirements.txt
firebase-admin==6.2.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
webdriver-manager==4.0.1
lxml==4.9.3
earthengine-api==0.1.364
